dashboard-api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
azure-identity==1.16.0
azure-digitaltwins-core==1.2.0
kubernetes==29.0.0